requirements1.txt
streamlit>=1.28
opencv-python>=4.7
numpy>=1.24
Pillow>=10.0
fpdf>=1.7
ultralytics>=8.0  # For YOLOv8 (optional)